+++ examples/render-frames/app/constants.ts
@@ -1,5 +1,4 @@
 export const HOST = process.env["NEXT_PUBLIC_HOST"];
 export const LOCAL_STORAGE_KEYS = {
   FARCASTER_USER: "farcasterUser",
-  // ... any other keys
 };
