+++ examples/render-frames/app/types/farcaster-user.ts
@@ -1,0 +1,10 @@
+export type FarcasterUser = {
+  signature: string;
+  publicKey: string;
+  privateKey: string;
+  deadline: number;
+  status: "approved" | "pending_approval";
+  signerApprovalUrl?: string;
+  token?: any;
+  fid?: number;
+};
